+++ source_echoargs_test.go
@@ -77,3 +77,40 @@ func TestEchoArgsWritesToPipelineStdout(t *testing.T) {
 	assert.Equal(t, expectedResult, pipeline.Pipe.Stdout.Strings())
 	assert.Equal(t, "", pipeline.Pipe.Stderr.String())
 }
+
+func TestEchoArgsWritesToTheTraceOutput(t *testing.T) {
+
+	// ----------------------------------------------------------------
+	// setup your test
+
+	expectedResult := `+ EchoArgs()
++ p.Stdout> hello world
++ p.Stdout> have a nice day
+`
+	dest := NewDest()
+	GetShellOptions().EnableTrace(dest)
+
+	testData := []string{"arg0 is not included", "hello world", "have a nice day"}
+
+	oldArgs := os.Args
+	os.Args = testData
+	// clean up after ourselves
+	defer func() {
+		os.Args = oldArgs
+	}()
+
+	pipeline := NewPipeline(
+		EchoArgs(),
+	)
+
+	// ----------------------------------------------------------------
+	// perform the change
+
+	pipeline.Exec()
+	actualResult := dest.String()
+
+	// ----------------------------------------------------------------
+	// test the results
+
+	assert.Equal(t, expectedResult, actualResult)
+}